refactor: commit 파일구조 정리
