--- src/main/java/com/leets/commitatobe/domain/commit/presentation/dto/request/CommitRequest.java
@@ -1,5 +1,0 @@
-package com.leets.commitatobe.domain.commit.presentation.dto.request;
-
-public record CommitRequest() {
-
-}
