+++ src/main/java/com/leets/commitatobe/domain/commit/repository/CommitRepository.java
@@ -1,4 +1,4 @@
-package com.leets.commitatobe.domain.commit.domain.repository;
+package com.leets.commitatobe.domain.commit.repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
